feat: create PromoCodes component
