+++ src/pages/Cart/Total/Total.tsx
@@ -1,25 +1,20 @@
-import { ReactElement, FormEvent, useState, useEffect } from 'react';
+import { ReactElement, useState, useEffect } from 'react';
 import { useSelector } from 'react-redux';
 import { useAppDispatch } from '../../../store/store';
-import { updateCartThunk, getDiscountsThunk } from '../../../store/cart/thunks';
+import { getDiscountsThunk } from '../../../store/cart/thunks';
 import { selectCartData } from '../../../store/cart/selectors';
 import { Button } from '../../../components/shared/Button';
-import { IPromoCode } from '../../../store/cart/types';
 import { formatPrice } from '../../../utils';
-import { ErrorMessage } from '../../../components/shared/ErrorMessage';
+import { PromoCodes } from '../PromoCodes';
 
 import styles from './total.module.scss';
 
 const DELIVERY_PRICE = 120;
-const CODE_ERROR = 'The code you entered is not valid.';
 
 export default function Total(): ReactElement {
-  const [code, setCode] = useState('');
-  const [isError, setIsError] = useState(false);
   const [isCheckoutPopup, setIsCheckoutPopup] = useState(false);
 
   const { basket, discounts } = useSelector(selectCartData);
-
   const dispatch = useAppDispatch();
 
   useEffect(() => {
@@ -35,55 +30,6 @@ export default function Total(): ReactElement {
       document.body.style.overflow = '';
     }
   }, [isCheckoutPopup]);
-
-  function getCodeName(codeId: string): string | undefined {
-    const code = discounts.find((item) => item.id === codeId);
-
-    return code?.code;
-  }
-
-  function applyCode(e: FormEvent<HTMLFormElement>): void {
-    e.preventDefault();
-
-    const isCodeValid = discounts.find((item) => item.code === code);
-
-    if (!isCodeValid) {
-      setIsError(true);
-      return;
-    }
-
-    dispatch(
-      updateCartThunk({
-        id: basket.id,
-        version: basket.version,
-        actions: [
-          {
-            action: 'addDiscountCode',
-            code: code,
-          },
-        ],
-      }),
-    );
-    setCode('');
-  }
-
-  function removeCode(item: IPromoCode): void {
-    dispatch(
-      updateCartThunk({
-        id: basket.id,
-        version: basket.version,
-        actions: [
-          {
-            action: 'removeDiscountCode',
-            discountCode: {
-              typeId: 'discount-code',
-              id: item.id,
-            },
-          },
-        ],
-      }),
-    );
-  }
 
   function countSubtotal(): number {
     const subtotal = basket.lineItems.reduce((acc, item) => {
@@ -127,39 +73,8 @@ export default function Total(): ReactElement {
         handleClick={(): void => setIsCheckoutPopup(true)}
       />
 
-      <div className={styles.promo}>
-        <h4 className={styles.promo__title}>Have a promo code?</h4>
-        <form className={styles.promo__form} onSubmit={(e): void => applyCode(e)} noValidate>
-          <input
-            className={styles.promo__form_input}
-            placeholder="Promo code"
-            value={code}
-            onInput={(event: React.FormEvent<HTMLInputElement>): void => setCode(event.currentTarget.value)}
-            onChange={(): void => setIsError(false)}
-          />
-          <Button type="submit" name="Apply" className={styles.promo__form_button} />
-        </form>
+      <PromoCodes basket={basket} discounts={discounts} dispatch={dispatch} />
 
-        {isError && <ErrorMessage text={CODE_ERROR} className={styles.promo__error} />}
-
-        {basket.codes.length > 0 && (
-          <ul className={styles.promo__codes}>
-            {basket.codes.map((item) => (
-              <li key={item.id} className={styles.code__item}>
-                {getCodeName(item.id) && (
-                  <span className={styles.code__item_label}>{getCodeName(item.id)} is applied</span>
-                )}
-                <Button
-                  type="button"
-                  name="Remove"
-                  className={styles.code__item_button}
-                  handleClick={(): void => removeCode(item)}
-                />
-              </li>
-            ))}
-          </ul>
-        )}
-      </div>
       {isCheckoutPopup && (
         <div className={styles.popup}>
           <div className={styles.popup__content}>
